packages/middleman_api/pyproject.toml
[tool.poetry]
name = "middleman-api"
version = "0.1.0"
description = ""
authors = ["MogboPython <mogbo18@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.2"
uvicorn = "^0.30.6"
motor = "^3.5.1"
pydantic = "^2.9.1"
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
python-jose = "^3.3.0"
pymongo = "^4.8.0"
shortuuid = "^1.0.13"
email-validator = "^2.2.0"
argon2-cffi = "^23.1.0"
httpx = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
